+++ src/infra/cron-jobs/ports/command.ts
@@ -1,0 +1,12 @@
+export interface RunnerArgs {
+  readonly uuid: string;
+  readonly instance: string;
+}
+
+interface Command {
+  run(args?: RunnerArgs): Promise<void>;
+  readonly name: string;
+  readonly schedule: string;
+}
+
+export default Command;