+++ src/main/app.ts
@@ -6,7 +6,6 @@ import { HttpServer } from '@/main/modules/http/http-server';
 import { Module } from '@/main/modules/modules';
 import { AMQPServer } from '@/main/modules/amqp/amqp-server';
 import { env } from '@/main/env';
-import { Worker } from './modules/worker/worker';
 import { CacheClient } from './modules/cache/cache-client';
 
 export class Application {
@@ -15,7 +14,6 @@ export class Application {
   protected loadModules(container: DependencyContainer): Module[] {
     return [
       new HttpServer(container),
-      new Worker(container),
       new AMQPServer(container, {
         protocol: env.rabbitMQProtocol,
         host: env.rabbitMQHost,